Attempt fix of FireBase error in BatteryStats line 59
Reason: Android does not allow Toast to be send on a thread with no Looper.
Either this is different in different Android versions, or there was
a race condition when the Toast was generated.

I have removed the Toast, it was just for debug info anyway.

In other locations I have modified the code to send Toast from UI thread.