Add display of DTC error codes, handle missing data by disconnecting BT
There seem to be some initial configuration of the communication
going on when connecting to BT, which does not kick in when simply
re-executing the init-sequence. So, as a last resort, if no data
are read from the car; disconnect blkuetooth.
If automatic reconnect is enabled, the issue may be gone after
connecting again.